Now connected to DB

+++ src/main/java/edu/wpi/fishfolk/ui/FlowerItem.java
@@ -3,11 +3,11 @@ package edu.wpi.fishfolk.ui;
 public class FlowerItem {
 
   public String itemName;
-  public int fullCost;
+  public double fullCost;
 
   public int amount;
 
-  public FlowerItem(String name, int price, int amount) {
+  public FlowerItem(String name, double price, int amount) {
     this.itemName = name;
     this.amount = amount;
     this.fullCost = price * amount;
